Initial p5.js project
The game as currently no goal it's hard to keep it minimalist with as less UI as possible (Even the score is against this game design philosophies)  i don't whant this game to become Snake, or Tron, so i need to let a new game play emerge from this node mechanic. i like the principle of cellular automata i would like the node logic to be similar in some sens. it could be that you die when all the node die, or are corrupted by opposite node a game play like that might not require a score or UI. which would keep the game more stylize.